requirements.txt
# 김엄마관리형독서실 학생회원관리 시스템 MVP - Django 백엔드 의존성

# Django 핵심 패키지
Django>=5.0.0,<5.1.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.0
django-extensions>=3.2.0

# 데이터베이스
psycopg2-binary>=2.9.7  # PostgreSQL
redis>=5.0.0
django-redis>=5.4.0

# 인증 및 보안
djangorestframework-simplejwt>=5.3.0
django-allauth>=0.57.0
cryptography>=41.0.0

# 실시간 통신
channels>=4.0.0
channels-redis>=4.1.0

# API 문서화
drf-spectacular>=0.26.5

# 유틸리티
python-decouple>=3.8  # 환경변수 관리
pillow>=10.0.0  # 이미지 처리
openpyxl>=3.1.0  # Excel 파일 처리
pandas>=2.1.0  # 데이터 처리 (구글시트 이관용)

# 개발 도구
pytest>=7.4.0
pytest-django>=4.5.2
black>=23.9.0  # 코드 포매터
flake8>=6.0.0  # 린터
factory-boy>=3.3.0  # 테스트 데이터 생성

# 배포 관련
gunicorn>=21.2.0
whitenoise>=6.6.0  # 정적 파일 서빙
